- Remove ServerFactory.

+++ java/org/apache/catalina/connector/MapperListener.java
@@ -29,7 +29,6 @@ import javax.management.ObjectName;
 import org.apache.catalina.ContainerEvent;
 import org.apache.catalina.ContainerListener;
 import org.apache.catalina.Host;
-import org.apache.catalina.ServerFactory;
 import org.apache.catalina.core.StandardContext;
 import org.apache.tomcat.util.buf.MessageBytes;
 import org.apache.tomcat.util.http.mapper.Mapper;
@@ -58,6 +57,11 @@ public class MapperListener
     protected Mapper mapper = null;
 
     /**
+     * Associated connector
+     */
+    protected Connector connector = null;
+
+    /**
      * MBean server.
      */
     protected MBeanServer mBeanServer = null;
@@ -71,7 +75,6 @@ public class MapperListener
 
     // It should be null - and fail if not set
     private String domain="*";
-    private String engine="*";
 
     // ----------------------------------------------------------- Constructors
 
@@ -79,8 +82,9 @@ public class MapperListener
     /**
      * Create mapper listener.
      */
-    public MapperListener(Mapper mapper) {
+    public MapperListener(Mapper mapper, Connector connector) {
         this.mapper = mapper;
+        this.connector = connector;
     }
 
 
@@ -92,14 +96,6 @@ public class MapperListener
 
     public void setDomain(String domain) {
         this.domain = domain;
-    }
-
-    public String getEngine() {
-        return engine;
-    }
-
-    public void setEngine(String engine) {
-        this.engine = engine;
     }
 
     /**
@@ -322,29 +318,32 @@ public class MapperListener
         throws Exception {
         String name=objectName.getKeyProperty("host");
         if( name != null ) {        
-            Host host = (Host) ServerFactory.getServer().findService(
-                    domain).getContainer().findChild(name);
+            Host host = (Host) connector.getService().getContainer().findChild(name);
             String[] aliases = host.findAliases();
             mapper.addHost(name, aliases, objectName);
             host.addContainerListener(this);
             if(log.isDebugEnabled())
                 log.debug(sm.getString
                      ("mapperListener.registerHost", name, domain));
-
         }
     }
 
 
-    /**
-     * Unregister host.
-     */
     private void unregisterHost(ObjectName objectName)
-        throws Exception {
+    throws Exception {
         String name=objectName.getKeyProperty("host");
-        mapper.removeHost(name);
-        if(log.isDebugEnabled())
-            log.debug(sm.getString
-                 ("mapperListener.unregisterHost", name, domain));
+        if( name != null ) { 
+            Host host =
+                (Host) connector.getService().getContainer().findChild(name);
+
+            mapper.removeHost(name);
+            if (host != null) {
+                host.removeContainerListener(this);
+            }
+            if(log.isDebugEnabled())
+                log.debug(sm.getString
+                        ("mapperListener.unregisterHost", name, domain));
+        }
     }
